Fix and add tests
