+++ packages/profile-sidebar/middleware.js
@@ -24,11 +24,9 @@ export default ({ dispatch, getState }) => next => action => {
     case actionTypes.PROFILE_ROUTE_LOADED: {
       const { selectedProfile = {} } = action;
       let profile = selectedProfile;
-      let profileFound = true;
 
       // force user redirection to valid profile if none exists
       if (!profile || !profile.id) {
-        profileFound = false;
         const { profileSidebar = {} } = getState();
         const { profiles } = profileSidebar;
         if (profiles && profiles.length > 0) {
@@ -41,12 +39,15 @@ export default ({ dispatch, getState }) => next => action => {
           profile,
         })
       );
+      break;
+    }
 
+    case actionTypes.SELECT_PROFILE: {
       dispatch(
         dataFetchActions.fetch({
           name: 'getCounts',
           args: {
-            profileId: profile.id,
+            profileId: action.profile.id,
           },
         })
       );
